// frontend/src/openapi/models/Behaviour.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Overbae API
 * Backend API for the Overclaw / Overmind platform.
 *
 * The version of the OpenAPI document: 2.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { BehaviourStatusEnum } from './BehaviourStatusEnum';
import {
    BehaviourStatusEnumFromJSON,
    BehaviourStatusEnumFromJSONTyped,
    BehaviourStatusEnumToJSON,
    BehaviourStatusEnumToJSONTyped,
} from './BehaviourStatusEnum';

/**
 *
 * @export
 * @interface Behaviour
 */
export interface Behaviour {
    /**
     *
     * @type {string}
     * @memberof Behaviour
     */
    readonly id: string;
    /**
     *
     * @type {string}
     * @memberof Behaviour
     */
    capability: string;
    /**
     *
     * @type {string}
     * @memberof Behaviour
     */
    key: string;
    /**
     *
     * @type {string}
     * @memberof Behaviour
     */
    displayName?: string;
    /**
     *
     * @type {string}
     * @memberof Behaviour
     */
    entryAnchor?: string;
    /**
     *
     * @type {BehaviourStatusEnum}
     * @memberof Behaviour
     */
    status?: BehaviourStatusEnum;
    /**
     *
     * @type {number}
     * @memberof Behaviour
     */
    readonly executionCount: number;
    /**
     *
     * @type {number}
     * @memberof Behaviour
     */
    readonly avgSuccessScore: number | null;
}



/**
 * Check if a given object implements the Behaviour interface.
 */
export function instanceOfBehaviour(value: object): value is Behaviour {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('capability' in value) || value['capability'] === undefined) return false;
    if (!('key' in value) || value['key'] === undefined) return false;
    if (!('executionCount' in value) || value['executionCount'] === undefined) return false;
    if (!('avgSuccessScore' in value) || value['avgSuccessScore'] === undefined) return false;
    return true;
}

export function BehaviourFromJSON(json: any): Behaviour {
    return BehaviourFromJSONTyped(json, false);
}

export function BehaviourFromJSONTyped(json: any, ignoreDiscriminator: boolean): Behaviour {
    if (json == null) {
        return json;
    }
    return {

        'id': json['id'],
        'capability': json['capability'],
        'key': json['key'],
        'displayName': json['display_name'] == null ? undefined : json['display_name'],
        'entryAnchor': json['entry_anchor'] == null ? undefined : json['entry_anchor'],
        'status': json['status'] == null ? undefined : BehaviourStatusEnumFromJSON(json['status']),
        'executionCount': json['execution_count'],
        'avgSuccessScore': json['avg_success_score'],
    };
}

export function BehaviourToJSON(json: any): Behaviour {
    return BehaviourToJSONTyped(json, false);
}

export function BehaviourToJSONTyped(value?: Omit<Behaviour, 'id'|'execution_count'|'avg_success_score'> | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'capability': value['capability'],
        'key': value['key'],
        'display_name': value['displayName'],
        'entry_anchor': value['entryAnchor'],
        'status': BehaviourStatusEnumToJSON(value['status']),
    };
}