// frontend/src/openapi/models/BehaviourEvaluator.ts
export interface BehaviourEvaluator {
    /**
     *
     * @type {string}
     * @memberof BehaviourEvaluator
     */
    readonly id: string;
    /**
     *
     * @type {string}
     * @memberof BehaviourEvaluator
     */
    name: string;
    /**
     *
     * @type {string}
     * @memberof BehaviourEvaluator
     */
    displayName?: string;
    /**
     *
     * @type {string}
     * @memberof BehaviourEvaluator
     */
    description?: string;
    /**
     *
     * @type {EvaluatorKindEnum}
     * @memberof BehaviourEvaluator
     */
    kind: EvaluatorKindEnum;
    /**
     *
     * @type {EvaluatorScopeEnum}
     * @memberof BehaviourEvaluator
     */
    scope?: EvaluatorScopeEnum;
    /**
     *
     * @type {ScoreDataTypeEnum}
     * @memberof BehaviourEvaluator
     */
    scoreType?: ScoreDataTypeEnum;
    /**
     *
     * @type {{ [key: string]: any; }}
     * @memberof BehaviourEvaluator
     */
    readonly behaviourBinding: { [key: string]: any; };
}